fix id duplication issue
